Fix bonus light window times
The current implementation shows a 2 hour long window starting at the most recent even-numbered hour in the user's local time, but this is incorrect since the bonus light windows are anchored to server time (UTC), so it will show the wrong value unless you are an even number of whole hours away from UTC.

Additionally, I suspect that the windows are not always two hours long, but instead the window rotates at 00:00 UTC, 02:00 UTC, 04:00 UTC, …, 22:00 UTC. This is almost the same thing, but it would show the wrong value for the 22:00 UTC - 00:00 UTC bonus window when there is a leap second, where this would say that the window is 22:00 - 23:59 instead of 22:00 - 00:00. Does this matter? Probably not, but since I was revisiting this I wanted to do it as correctly as I could.

This new code uses window start times instead of durations and calculates everything in UTC before converting it to the user's local time.